Implement feat Process list pausing
This change sets a keybinding 'F' that pauses the processes list.
It doesn't stop collection.
It stops new processes from being added and dead processes from being removed.
It also disables sorting except when the sorting mode is changed or when switching between normal and tree view modes.
While paused the user can still scroll through the processes list.
Detailed view can still be opened and it displays current information even when the process is dead.
When pausing is enabled it displays a banner on the last line of the processes list.

the BSDs need testing still.